TC:Almost workable version
Note: This is a WiP version and should NOT be reviewed or merged.

This version is used to run the correctness test:
20180926-180645-mengxu-9c5c19ef6e002bd7
which has 2 failures out of 55751 runs in 4 hours

This version seems working well in most cases.
It can serve as a base point for the further bug fix.

I checked 1 ouf the 2 failures. It fails because of BestTeamStuck.
This is because all servers are unhealthy and the req.req.completeSources is empty.

Q: How should we handle the situation?
Q: Maybe the test workload is not correct?
In such a situation, the DB should be marked as unhealthy and we should let the system operator to handle this,
because it is better to "crash" a system instead of leaving the system in an unknown state. (Availability vs. Reliability)

Meng Xu