TC:correctly count the number of teams a machine or server has
A machine (or server) can have a machine team with incorrect size.
When we find the least used machine or server, we should count only the number of teams with the correct size.
Previously, we also count the teams with incorrect size.

This will exclude some machine (server) that belong to incorrect-size team but should be used to created new teams.
This situation does not happen often, but when it happens, we will have BestTeamStuck error.