decode range and log file succeeed
Successfully decode range and log file
Note: files are encoded in blocks. We need to decode each file block by block.

Note: The value in the log file encodes the mutaiton type. The key in the log file encodes the version of the mutation.
Need to further decode the k-v in the log file to get the k-v that will be applied to DB for restoration.

Meng Xu