FastRestore:Master:Debug:Add delay in each phase
I suspect the commands in previous phase may be received in the current phase;
When it happens, the local states on Loader and Applier may be incorrect, say
we may get extra mutations.

The design did not consider this situation.

If adding delay between phases can reduce the error rate, the abboe suspection is valid.
We should add new command to move nodes to each phase and filter out commands from previous phases.