FastRestore:Add AtomicAdd test
AtomicAdd continuously increase a counter from 0;
For each counter value, it writes a single key generated from the value.

The result should be a key space with continuous key whose value is
equal to its key's counter value.

Example, (ops/1/2, 2), where 1 is the clientId and 2 is the counter value.
The existance of this kv means we also have ops/1/1 and ops/1/0 key.